Fixed a bug in the camp calculation

+++ examples/SimpleAxis.py
@@ -20,7 +20,7 @@ def wait_for_responses(axis, responses_to_check, msg):
 steps_per_rev = int(input("How many steps per revolution [1600]? ") or "1600")
 total_revolutions = int(input("How many total revolutions [50]? ") or "50")
 rpm = int(input("How many RPMs [500]? ") or "500")
-rdc = rpm / 10      # Calculate ramp divide. 10% of rpm seems to work at least up to 1500 rpm
+rdc = int(rpm / 10)      # Calculate ramp divide. 10% of rpm seems to work at least up to 1500 rpm
 ramp_divide = int(input(f"Ramp divide [{rdc}]? ") or f"{rdc}")
 ramp_pause = int(input("Ramp pause [10]? ") or "10")
 direct = input("Direction (Forward = F, Reverse = R) [F]? ") or "F"
